# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  BadRequestError:
    properties:
      error:
        example: Bad Request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  DeleteUserSuccess:
    properties:
      data:
        example: User deleted successfully.
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  InternalServerError:
    properties:
      error:
        example: Internal Server Error
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  Success-DeleteUserSuccess:
    properties:
      data:
        $ref: '#/definitions/DeleteUserSuccess'
      success:
        example: true
        type: boolean
    type: object
...
  Success-UpdateUserSuccess:
    properties:
      data:
        $ref: '#/definitions/UpdateUserSuccess'
      success:
        example: true
        type: boolean
    type: object
  Success-UserResponse:
    properties:
      data:
        $ref: '#/definitions/UserResponse'
      success:
        example: true
        type: boolean
    type: object
  Success-array_UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/UserResponse'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  UnauthorizedError:
    properties:
      error:
        example: Unauthorized
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  UpdateUserSuccess:
    properties:
      data:
        example: User upadated successfully.
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  UserRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        maxLength: 50
        minLength: 2
        type: string
      phone:
        type: string
      role_id:
        type: string
    type: object
  UserResponse:
    properties:
      email:
        example: john.doe@gmail.com
        type: string
      first_name:
        example: John
        type: string
      full_name:
        example: John Doe
        type: string
      last_name:
        example: Doe
        type: string
      phone:
        example: "1234567890"
        type: string
      role_id:
        example: 97d699c0-24ff-48dc-b64a-c29353fa8865
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: jaiswal.vaibhavkumar45@gmail.com
    name: API Support
  description: API for managing products and orders.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Returns a list of users with optional filter/query parameters.
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/Success-array_UserResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/BadRequestError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Get Users with Filters
      tags:
      - User
  /user/{user_id}:
    delete:
      description: Deletes a user identified by the provided ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/Success-DeleteUserSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - User
    get:
      description: Retrieves a user's details by their unique ID.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data fetched successfully
          schema:
            $ref: '#/definitions/Success-UserResponse'
        "400":
          description: Invalid ID or user not found
          schema:
            $ref: '#/definitions/BadRequestError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Partially update user fields using PATCH request
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Partial Update User Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User partially updated successfully
          schema:
            $ref: '#/definitions/Success-UpdateUserSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Partially update user by ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user details based on the provided ID and request body
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Update User Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/Success-UpdateUserSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Update user by ID
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"